parisc: Fix usage of local tlb purges in pacache.S
Allow to build a 32-bit kernel which is tuned for P8X00 CPUs
(CONFIG_PA20=y), which still is able to boot on PA1.x machines.

Since 32-bit CPUs don't support the "local" flag of the "pdtb,l"
instruction, reorganize the flush instructions so that the alternative
patching routine can patch in the local flag at runtime when runing on
PA8X00 CPUs.

Note that there is an additonal check in the alternative patching code,
which prevents enabling the local flag if the kernel doesn't run on a
PA2.0 CPU.  So, the ALTERNATIVE() macros below will only be applied at
runtime on PA2.0 machines.

Follow up code still needs to address the bus serialization issue when
pdtlb is used on an SMP machine.

Signed-off-by: Helge Deller <deller@gmx.de>
Cc: stable@vger.kernel.org