parisc: Fix prefetch() and prefetchw() on 32-bit CPUs
Main motivation for this patch is to allow to build a 32-bit kernel
tuned for P8X00 CPUs while still being able to boot on PA1.x machines.

For that make sure that no 64-bit instructions are used on a 32-bit
kernel and ignore accesses to page 0 due to the PA7300LC prefetch bug
(see comments at top of this header file).

A 64-bit build will instead use the __builtin_prefetch() provided by the
compiler and in include/linux/prefetch.h.

Signed-off-by: Helge Deller <deller@gmx.de>
Cc: stable@vger.kernel.org