parisc: Add unlocked functions for pdc_iodc_getc() and pdc_iodc_print()
The built-in kernel debugger (kgdb) deadlocks if we use spinlocks (which
should protect against concurrent firmware calls) to get keyboard input
or to output to the console.  Since we are in debugger mode anyway, we
can simply avoid using this protection mechanism, which is why this
patch adds unlocked keyboard input/console output functions which will
be used by a follow-up patch.

In addition save some static bytes by reusing pdc_result[] as
iodc_retbuf[] and by reducing the memory footprint of iodc_dbuf[].

Signed-off-by: Helge Deller <deller@gmx.de>